requirements.txt
fastapi
uvicorn
python-docx
pymupdf
pytesseract
pillow